pyproject.toml
[project]
name = "nn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jax-metal==0.1.1",
    "jax==0.5.0",
    "numpy>=2.2.4",
    "pillow>=11.2.1",
    "pygame-ce>=2.5.3",
]

[tool.ruff]
indent-width = 2
